Refactor Cannery wrapper behind protocol session
Keep the Phase 0 JSON contract isolated so generated protocol consumers can replace it without changing CLI commands or process supervision.
