+++ truss/cli/cannery/__init__.py
@@ -1,0 +1,1 @@
+"""Truss integration boundary for the Cannery client."""
